zpodapi/pyproject.toml
[tool.poetry]
name = "zpodapi"
version = "0.2.3"
description = ""
authors = ["Kelby Valenti <kvalenti@vmware.com>", "Timo Sugliani <tsugliani@vmware.com>"]
readme = "README.md"

[tool.poetry.dependencies]
alembic = "1.13.1"
email-validator = "2.1.0.post1"
fastapi = "0.109.0"
gitpython = "3.1.41"
gunicorn = "21.2.0"
httpx = "0.26.0"
ipython = "^8.20.0"
passlib = "1.7.4"
psycopg2 = "2.9.9"
pydantic-settings = "2.1.0"
python = "3.12.1"
python-dotenv = "1.0.1"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
python-multipart = "0.0.6"
pyvmomi = "8.0.2.0.1"
rich = "13.7.0"
sqlmodel = "0.0.14"
tenacity = "8.2.3"
uvicorn = "0.27.0.post1"

[tool.poetry.group.dev.dependencies]
coverage = "^7.1.0"
debugpy = "^1.6.6"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-xdist = "^3.2.0"
ruff = "^0.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
